Updated CSS
All works but is almost exactly at 100kb in size meaning its right on the file limit. (I removed some redundant Emotes for now. If you want to optimize it more and bring it down to around the old subreddit style it would be great. (Old one is at 95KB)